+++ extension/CHANGELOG.md
@@ -1,8 +1,6 @@
 # Change Log
 
-## 0.9.0
-- Support for tracking channels per repository/extension.
-
 ## Unreleased
 
 - Initial release
+  - Support for tracking channels per repository/extension.